+++ App.tsx
@@ -1,6 +1,21 @@
 import React from 'react'
-import { View } from 'react-native'
+import { View, StyleSheet } from 'react-native'
+import DropdownAlert from 'react-native-dropdownalert'
+import AppAlertManager from './src/helpers/AppAlertManager'
 
-const App = () => <View />
+const App = () => (
+    <View style={styles.container}>
+        <DropdownAlert
+            closeInterval={2500}
+            ref={(ref: DropdownAlert) => AppAlertManager.setDropdown(ref)}
+        />
+    </View>
+)
+
+const styles = StyleSheet.create({
+    container: {
+        flex: 1,
+    },
+})
 
 export default App